Fix lastModifiedAt time

+++ host-factory-generator/src/main/java/react4j/hfg/Fetch.java
@@ -122,7 +122,7 @@ public final class Fetch
   @SuppressWarnings( "SameParameterValue" )
   @Nonnull
   private static Path fetchData( @Nonnull final String url,
-                                 final int lastModifiedAt,
+                                 final long lastModifiedAt,
                                  @Nonnull final String targetFileName )
     throws FetchException
   {
